refactor: organized files into folders
