+++ src/js/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import 'src/css/index.css';
+import App from 'src/js/App';
+import reportWebVitals from 'src/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
